Moved the dialog around
