+++ src/dialogs/addImageDialog.js
